Revert "Merge branch 'Experiment' into clone-main"
This reverts commit 205afa1fc82f8d1a3d5de2d9cf9e6fb04c0f8dc9, reversing
changes made to dfb85ac0e4b8cd1d7bfe0dcbe2aface0f78f8593.

+++ Assets/Scripts/Entity/StatusEffect/Buff/CRUP.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewCRUP", menuName = "StatusEffect/CRUP")]
 public class CRUP : StatusEffect
 {
-    public float PercentAmount;
+    public int PercentAmount;
 
     public override void OnApply(CharacterInBattle target)
     {
